refactor(ft_itoa.c): Simplify conversion to char array
The function ft_get_len has been removed and its logic has been incorporated into the ft_itoa function. The conversion of the integer to a char array has also been simplified by using an array of digits instead of manually adding the ascii value of '0' to the result of the modulo operation.