feat: add ft_atof function to libft library
This commit adds a new function ft_atof to the libft library. The function converts a string to a double-precision floating-point number. The function is implemented in the ft_atof.c file. The function is added to the Makefile and the function prototype is added to the libft.h header file.